src/foundry-sdk-python/deploy-pumps-agent/requirements.txt
azure-ai-projects
azure-ai-agents
openai
azure-identity
azure-mgmt-cognitiveservices
azure-mgmt-authorization
azure-search-documents
pytest
python-dotenv
requests